requirements.txt
flywheel-sdk>=18.0.0
python-dotenv>=1.0.0
